test: isolate report fallback evidence
