+++ e2e/bug-report.spec.ts
@@ -107,36 +107,6 @@ test("failed submission preserves the draft in the GitHub fallback", async ({ pa
   expect(url.searchParams.get("title")).toBe("Keep my feature title");
   expect(url.searchParams.get("body")).toContain("Keep my typed description 🐛");
   expect(url.searchParams.get("labels")).toBe("enhancement");
-
-  const github = await page.context().newPage();
-  await github.goto(url.toString(), { waitUntil: "domcontentloaded" });
-  await expect(github).toHaveURL(/github\.com\/(?:login|xnmp\/tauri-explorer\/issues\/new)/);
-  await github.screenshot({ path: evidencePath("ac-1-github-fallback.png") });
-  await github.close();
-});
-
-test("report fallback error toast has an opaque surface", async ({ page }) => {
-  await page.goto("/");
-  await waitForEntries(page);
-  await page.evaluate(() => localStorage.setItem("mock-report-error", "network_unreachable"));
-
-  await runPaletteCommand(page, "Report a Bug");
-  const dialog = page.getByRole("dialog", { name: /report a bug/i });
-  await dialog.getByLabel("Title").fill("Opaque fallback toast");
-  await dialog.getByLabel("Description").fill("Keep this report in GitHub.");
-  await dialog.getByRole("button", { name: "Submit" }).click();
-
-  const toast = page.locator(".toast.error");
-  await expect(toast).toContainText("Could not submit in-app — opening GitHub instead");
-  await expect.poll(() => toast.evaluate((element) => {
-    const style = getComputedStyle(element);
-    const channels = style.backgroundColor.match(/[\d.]+/g) ?? [];
-    return {
-      alpha: style.backgroundColor.startsWith("rgba") ? Number(channels[3]) : 1,
-      backgroundImage: style.backgroundImage,
-    };
-  })).toEqual({ alpha: 1, backgroundImage: "none" });
-  await toast.screenshot({ path: evidencePath("ac-2-report-fallback-toast.png") });
 });
 
 test("image picker shows named previews and lets the user remove an attachment", async ({ page }) => {
